+++ context_test.go
@@ -345,3 +345,21 @@ func TestContext_XML(t *testing.T) {
 		assert.Equal(t, w.Body.String(), test.body, "resposne body does not match")
 	}
 }
+
+func TestContext_HTML(t *testing.T) {
+	tests := []struct {
+		code int
+		s    string
+	}{
+		{200, "<html><body>foobar</body></html>"},
+		{500, "<html><body>error</body></html>"},
+	}
+	for _, test := range tests {
+		w := httptest.NewRecorder()
+		ctx := newContext(w, nil)
+		ctx.HTML(test.code, test.s)
+		assert.Equal(t, test.code, w.Code, "status code does not match")
+		assert.Equal(t, w.Header().Get("Content-Type"), "text/html; charset=utf-8", "content type does not match")
+		assert.Equal(t, w.Body.String(), test.s, "resposne body does not match")
+	}
+}