+++ CHANGELOG.md
@@ -3,10 +3,11 @@ Change Log
 
 under development
 -----------------
-- Add `Context.JSON` to sends JSON response.
-- Add `Context.String` to sends string response.
-- Add `Context.XML` to sends XML response.
+- Add `Context.JSON` to send JSON response.
+- Add `Context.String` to send string response.
+- Add `Context.XML` to send XML response.
 - `Content.SetContentTypeJSON` and `Content.SetContentTypeXML` append `charset=utf-8` to content type header.
+- Add `Context.HTML` to send HTML response.
 
 v1.8.1 April 2, 2020
 --------------------